git-svn-id: svn://svn.icms.temple.edu/lammps-ro/trunk@8194 f3b2605a-c512-4ea7-a41b-209d697bcdaa

+++ src/MOLECULE/angle_table.cpp
@@ -293,7 +293,7 @@ void AngleTable::read_restart(FILE *fp)
     fread(&tabstyle,sizeof(int),1,fp);
     fread(&tablength,sizeof(int),1,fp);
   }
-  MPI_Bcast(&tabstyle,1,MPI_DOUBLE,0,world);
+  MPI_Bcast(&tabstyle,1,MPI_INT,0,world);
   MPI_Bcast(&tablength,1,MPI_INT,0,world);
 
   allocate();
